Initialize min codebase
This sets up the codebase with the normal necessities: Code of Conduct,
Licenses and a readme. It also sets up the core logic to allow min to be
able to take async route handler functions as well as a basic example
for how it's currently used. From here we need to start building out
tooling to determine what conventions we want to follow as well as
packing in everything for a batteries included experience. Ergonomics is
key to pleasant experience. If someone can pick up this framework with
only a cursory knowledge of Rust and can start building applications
then it's a success.