Simplify router logic and add all HTTP verbs to it
During the prototyping phase we implemented a `Service` by hand in order
to get the hyper server. However, this was unnecessary and the type and
it's impls could be removed leaving just the `Router` type to which we
have also added the last functions needed in order to add routes for
each of the HTTP verb types defined in IETF RFC 7231.